third_party/WebKit/LayoutTests/svg/custom/svg-parse-overflow-1-expected.txt

CONSOLE ERROR: line 9: Error: <foreignObject> attribute y: Expected length, "1050000000000000…".
